deprecate current CD
